Extract the media-session cluster from daemon engine.rs
engine.rs sat one line under its ratchet cap (1875 vs 1874+grace), so
any work in the file was about to be blocked. Move the OS media-session
methods (set_media_art, handle_media, capability probes,
media_set_rating, media_snapshot) into engine/media_session.rs as a
pure code move — parity-sensitive logic is byte-identical, verified by
adversarial review against main. media_can_seek and
current_is_radio_stream become pub(super) for their existing
engine-submodule callers. Baseline row re-blessed 1874 -> 1576.

runner.rs (1536 vs cap 1532, inside grace) is deliberately left alone:
its extractable helpers interleave with the IME-scrub burst contract
frozen in PR #83, and it blocks nothing yet.
