+++ scripts/size-baseline.tsv
@@ -1,4 +1,4 @@
-1874	src/daemon/engine.rs
+1576	src/daemon/engine.rs
 1638	src/app/local.rs
 1591	src/player/lifetime.rs
 1573	src/ui/anim/canvas.rs
